convert/requirements.txt
torch>=2.0
torchvision
timm
numpy
Pillow
onnx
onnxsim
onnx2tf
tensorflow>=2.15,<2.17
tf-keras
